requirements.txt
attrs==18.1.0
psycopg2==2.7.5
python-dateutil==2.7.3
requests==2.19.1
SQLAlchemy==1.2.9

# Transitive
certifi==2018.4.16
chardet==3.0.4
idna==2.7
six==1.11.0
urllib3==1.23
